+++ src/pages/landing/Landing.jsx
@@ -1,6 +1,11 @@
 import './landing.css';
-import {useState} from 'react';
+import axios from 'axios';
+import {useState, useEffect} from 'react';
 import CatCarousel from '../../components/carousels/CatCarousel';
+import PostsCarousel from '../../components/carousels/postsCarousel';
+import ArrowBackIosIcon from '@mui/icons-material/ArrowBackIos';
+import ArrowForwardIosIcon from '@mui/icons-material/ArrowForwardIos';
+
 // import ToolsCard from '../../components/tools/ToolsCard';
 import { Card } from '@mui/joy';
 
@@ -9,6 +14,43 @@ import Tools from '../../components/tools/Tools';
 import Posts from '../../components/posts/Posts';
 
 export default function Landing({setModal}) {
+  const API = import.meta.env.VITE_REACT_APP_API_URL;
+  const [posts, setposts] = useState([]);
+  const [currentPost, setCurrentPost] = useState(0);
+  const [visiblePosts, setVisiblePosts] = useState([]);
+
+  useEffect(() => {
+    const getPosts = () => {
+    axios.get(`${API}/posts`)
+    .then((response) => {
+      const allPosts = response.data;
+      const theVisiblePosts = [
+        allPosts[(currentPost - 1 + allPosts.length) % allPosts.length],
+        allPosts[currentPost],
+        allPosts[(currentPost + 1) % allPosts.length],
+        allPosts[(currentPost + 2) % allPosts.length],
+        allPosts[(currentPost + 3) % allPosts.length],
+      ];
+      setVisiblePosts(theVisiblePosts);
+      setposts(response.data);
+    })
+    .catch(error => console.error('catch', error))
+  }
+  getPosts();
+}, [currentPost, API]);
+
+  function prevSlide() {
+    setCurrentPost(prevPost =>
+      prevPost === 0 ? posts.length - 1 : prevPost - 1
+    );
+  } 
+
+  function nextSlide() {
+    setCurrentPost(prevPost =>
+      prevPost === posts.length - 1 ? 0 : prevPost + 1
+    );
+  } 
+
   return (
     <div>
       <header>
@@ -42,8 +84,69 @@ export default function Landing({setModal}) {
         <br />
       </header>
       <br />
-      <Posts />
-      <Tools />
+      <main>
+      <h3> Top Categories </h3>
+      <br />
+      <div className='top-category-1'>
+      <h4 className='main-h4'> Painting </h4>
+      <div className='posts-slider-container'>
+      <button className='arrow' onClick={prevSlide}>{' '} <ArrowBackIosIcon />{' '} </button>
+      { 
+        visiblePosts.map((post, i) => { 
+          //console.log(post)
+          if (post.category === 'Paint') {
+          return (
+            <>
+            <PostsCarousel post={post} i={i} />
+            </>
+          )
+          }
+        })
+      }
+       <button className='arrow' onClick={nextSlide} >
+        {' '}
+        <ArrowForwardIosIcon />{' '}
+      </button>
+      </div>
+      </div>
+      <div className='top-category-2'>
+        <h4 className='main-h4'> Sculpting </h4>
+        <div className='posts-slider-container'>
+        <button className='arrow' onClick={prevSlide}>{' '} <ArrowBackIosIcon />{' '} </button>
+        { 
+        visiblePosts.map((post, i) => { 
+          //console.log(post)
+          if (post.category === 'Sculpt') {
+          return (
+            <>
+            <PostsCarousel post={post} i={i} />
+            </>
+          )
+          }
+        })
+        }
+        </div>
+      </div>
+      <div className='top-category-3'>
+        <h4 className='main-h4'> Sketching </h4>
+        <div className='posts-slider-container'>
+        <button className='arrow' onClick={prevSlide}>{' '} <ArrowBackIosIcon />{' '} </button>
+        { 
+        visiblePosts.map((post, i) => { 
+          //console.log(post)
+          if (post.category === 'Sketch') {
+          return (
+            <>
+            <PostsCarousel post={post} i={i} />
+            </>
+          )
+          }
+        })
+        }
+        </div>
+      </div>
+      </main>
+      {/* <Tools /> */}
     </div>
   );
 }
