+++ src/pages/landing/landing.css
@@ -8,6 +8,26 @@
   margin-bottom: 40px;
 }
 
+.posts-slider-container {
+  position: relative;
+  display: flex;
+  flex-direction: row;
+  gap: 19px;
+  justify-content: center;
+  align-items: center;
+  width: 90%;
+  margin: 0 auto;
+}
+
+.main-h4 {
+display: flex;
+margin-left: 40px;
+font-family: 'Roboto';
+color: #1a237e;
+font-weight: lighter;
+font-size: large;
+}
+
 .landing-action {
   display: grid;
   grid-template-columns: 1fr 2fr;
